[Silver III] Title: 주유소, Time: 20 ms, Memory: 2804 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -12,7 +12,7 @@ int main() {
     for (int i=0; i<n; i++) cin >> cos[i];
     for (int i=0; i<n; i++) {
         c=min(c, cos[i]);
-        r+=c*dis[i];
+        r+=(int64_t)c*dis[i];
     }
     cout << r;
 }
